refactor: use clap for parse command line
