+++ src-tauri/src/main.rs
@@ -5,10 +5,14 @@
 mod webview2;
 
 fn main() {
-    if mxu_lib::commands::system::has_help_flag() {
-        mxu_lib::commands::system::print_cli_help_text();
-        std::process::exit(0);
-    }
+    #[cfg(windows)]
+    attach_parent_console_for_cli();
+
+    // Parse CLI args; clap handles --help display and exit automatically
+    let _cli = mxu_lib::commands::system::init_cli();
+
+    #[cfg(windows)]
+    detach_parent_console_for_cli();
 
     #[cfg(target_os = "windows")]
     {
@@ -75,4 +79,30 @@ fn main() {
     }
 
     mxu_lib::run()
+}
+
+#[cfg(windows)]
+fn attach_parent_console_for_cli() {
+    extern "system" {
+        fn AttachConsole(dw_process_id: u32) -> i32;
+    }
+
+    const ATTACH_PARENT_PROCESS: u32 = 0xFFFF_FFFF;
+
+    // GUI subsystem builds do not auto-attach to the invoking terminal.
+    // Ignore failure: redirected stdout or double-click launches should still fall through.
+    unsafe {
+        AttachConsole(ATTACH_PARENT_PROCESS);
+    }
+}
+
+#[cfg(windows)]
+fn detach_parent_console_for_cli() {
+    extern "system" {
+        fn FreeConsole() -> i32;
+    }
+
+    unsafe {
+        FreeConsole();
+    }
 }
